requirements.txt
Pillow>=10.0.0
numpy>=1.24.0,<2.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
pandas>=2.0.0
nibabel>=5.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0
xgboost>=2.0.0
tensorflow>=2.13.0,<2.16.0